set up react router for app, refactor part of authentication code into callback route, create placeholder dashboard page
