+++ frontend/src/App.js
@@ -1,10 +1,23 @@
 import './App.css';
-import SpotifyButton from "./SpotifyButton"
+import {Routes, Route} from "react-router-dom";
+import Login from "./Login"
+import Layout from "./Layout"
+import Dashboard from "./Dashboard";
+import Callback from "./Callback"
 
 function App() {
-  return (
-    <SpotifyButton/>
-  );
+    return (
+        <div>
+            <p>hello world!</p>
+            <Routes>
+                <Route path="/" element={<Layout/>}>
+                    <Route index element={<Login/>}/>
+                    <Route path="dashboard" element={<Dashboard/>}/>
+                    <Route path="callback" element={<Callback/>}/>
+                </Route>
+            </Routes>
+        </div>
+    );
 }
 
 export default App;
