+++ frontend/src/Callback.js
@@ -1,0 +1,81 @@
+import {useEffect} from "react";
+
+const redirectUri = process.env.REACT_APP_REDIRECT_URI;
+const clientId = process.env.REACT_APP_CLIENT_ID;
+
+async function getApiToken(accessToken, refreshToken, expiresIn) {
+    const url = "http://localhost:8000/api/authenticate"
+
+    const payload = {
+      method: 'POST',
+      headers: {
+        'Content-Type': 'application/x-www-form-urlencoded',
+      },
+      body: new URLSearchParams({
+          'access_token': accessToken,
+          'refresh_token': refreshToken,
+          'expires_in': expiresIn
+      }),
+    }
+
+    const body = await fetch(url, payload);
+    if (!body.ok) {
+        console.error("Unable to authenticate user");
+        return Promise.reject()
+    }
+
+    const response = await body.json();
+    const token = response.auth_token;
+
+    console.log("Token acquired: " + token)
+    localStorage.setItem("token", token);
+}
+
+const getToken = async (code) => {
+    const url = "https://accounts.spotify.com/api/token";
+    // stored in the previous step
+    let codeVerifier = localStorage.getItem('code_verifier');
+
+    const payload = {
+      method: 'POST',
+      headers: {
+        'Content-Type': 'application/x-www-form-urlencoded',
+      },
+      body: new URLSearchParams({
+        client_id: clientId,
+        grant_type: 'authorization_code',
+        code,
+        redirect_uri: redirectUri,
+        code_verifier: codeVerifier,
+      }),
+    }
+
+    const body = await fetch(url, payload);
+
+    if (!body.ok) {
+        console.error("Unable to fetch token from Spotify API");
+        return Promise.reject();
+    }
+
+    const response = await body.json();
+
+    const accessToken = response.access_token;
+    const refreshToken = response.refresh_token;
+    const expiresIn = response.expires_in;
+
+    await getApiToken(accessToken, refreshToken, expiresIn);
+}
+
+function Callback() {
+    useEffect(() => {
+      const urlParams = new URLSearchParams(window.location.search);
+      let code = urlParams.get('code');
+      if (code) {
+          getToken(code);
+      }
+    }, []);
+
+    return (<div>Please wait...</div>);
+}
+
+export default Callback;
