+++ frontend/src/Dashboard.js
@@ -1,0 +1,9 @@
+function Dashboard() {
+    return (
+        <div>
+            <p>Welcome to the dashboard.</p>
+        </div>
+    )
+}
+
+export default Dashboard;
