+++ frontend/src/Login.js
@@ -1,0 +1,8 @@
+import SpotifyButton from "./SpotifyButton"
+function Login() {
+    return (
+        <SpotifyButton/>
+    );
+}
+
+export default Login;
